requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
langgraph==1.0.0a3
langchain==1.0.0a5
pydantic

# Google AI Studio
langchain-google-genai

# Etl com excel
pandas
openpyxl